+++ CHANGELOG.md
@@ -19,6 +19,7 @@ Test and beta releases will have experimental functions enabled by default.
 * Sharper check marks, thanks @[jpeg729](https://github.com/jpeg729)
 * Toggle (clear/set) selected applications
 * Fetch/submit selected applications
+* Import selected applications
 * Updated simplified Chinese translation
 * Updated traditional Chinese translation
 * Updated French translation